+++ crates/bevy_core_pipeline/src/core_3d/mod.rs
@@ -1,5 +1,6 @@
 mod camera_3d;
-mod main_pass_3d_node;
+mod main_opaque_pass_3d_node;
+mod main_transparent_pass_3d_node;
 
 pub mod graph {
     pub const NAME: &str = "core_3d";
@@ -9,7 +10,10 @@ pub mod graph {
     pub mod node {
         pub const MSAA_WRITEBACK: &str = "msaa_writeback";
         pub const PREPASS: &str = "prepass";
-        pub const MAIN_PASS: &str = "main_pass";
+        pub const START_MAIN_PASS: &str = "start_main_pass";
+        pub const MAIN_OPAQUE_PASS: &str = "main_opaque_pass";
+        pub const MAIN_TRANSPARENT_PASS: &str = "main_transparent_pass";
+        pub const END_MAIN_PASS: &str = "end_main_pass";
         pub const BLOOM: &str = "bloom";
         pub const TONEMAPPING: &str = "tonemapping";
         pub const FXAA: &str = "fxaa";
@@ -21,7 +25,8 @@ pub mod graph {
 use std::cmp::Reverse;
 
 pub use camera_3d::*;
-pub use main_pass_3d_node::*;
+pub use main_opaque_pass_3d_node::*;
+pub use main_transparent_pass_3d_node::*;
 
 use bevy_app::{App, Plugin};
 use bevy_ecs::prelude::*;
@@ -82,20 +87,33 @@ impl Plugin for Core3dPlugin {
             );
 
         let prepass_node = PrepassNode::new(&mut render_app.world);
-        let pass_node_3d = MainPass3dNode::new(&mut render_app.world);
+        let opaque_node_3d = MainOpaquePass3dNode::new(&mut render_app.world);
+        let transparent_node_3d = MainTransparentPass3dNode::new(&mut render_app.world);
         let tonemapping = TonemappingNode::new(&mut render_app.world);
         let upscaling = UpscalingNode::new(&mut render_app.world);
         let mut graph = render_app.world.resource_mut::<RenderGraph>();
 
         let mut draw_3d_graph = RenderGraph::default();
         draw_3d_graph.add_node(graph::node::PREPASS, prepass_node);
-        draw_3d_graph.add_node(graph::node::MAIN_PASS, pass_node_3d);
+        draw_3d_graph.add_node(graph::node::START_MAIN_PASS, EmptyNode);
+        draw_3d_graph.add_node(graph::node::MAIN_OPAQUE_PASS, opaque_node_3d);
+        draw_3d_graph.add_node(graph::node::MAIN_TRANSPARENT_PASS, transparent_node_3d);
+        draw_3d_graph.add_node(graph::node::END_MAIN_PASS, EmptyNode);
         draw_3d_graph.add_node(graph::node::TONEMAPPING, tonemapping);
         draw_3d_graph.add_node(graph::node::END_MAIN_PASS_POST_PROCESSING, EmptyNode);
         draw_3d_graph.add_node(graph::node::UPSCALING, upscaling);
 
-        draw_3d_graph.add_node_edge(graph::node::PREPASS, graph::node::MAIN_PASS);
-        draw_3d_graph.add_node_edge(graph::node::MAIN_PASS, graph::node::TONEMAPPING);
+        draw_3d_graph.add_node_edge(graph::node::PREPASS, graph::node::START_MAIN_PASS);
+        draw_3d_graph.add_node_edge(graph::node::START_MAIN_PASS, graph::node::MAIN_OPAQUE_PASS);
+        draw_3d_graph.add_node_edge(
+            graph::node::MAIN_OPAQUE_PASS,
+            graph::node::MAIN_TRANSPARENT_PASS,
+        );
+        draw_3d_graph.add_node_edge(
+            graph::node::MAIN_TRANSPARENT_PASS,
+            graph::node::END_MAIN_PASS,
+        );
+        draw_3d_graph.add_node_edge(graph::node::END_MAIN_PASS, graph::node::TONEMAPPING);
         draw_3d_graph.add_node_edge(
             graph::node::TONEMAPPING,
             graph::node::END_MAIN_PASS_POST_PROCESSING,
