+++ crates/bevy_pbr/src/lib.rs
@@ -292,7 +292,7 @@ impl Plugin for PbrPlugin {
         draw_3d_graph.add_node(draw_3d_graph::node::SHADOW_PASS, shadow_pass_node);
         draw_3d_graph.add_node_edge(
             draw_3d_graph::node::SHADOW_PASS,
-            bevy_core_pipeline::core_3d::graph::node::MAIN_PASS,
+            bevy_core_pipeline::core_3d::graph::node::START_MAIN_PASS,
         );
     }
 }
